可視化ツールを作成

+++ n_d_tensor_with_ghost_cell.h
@@ -14,7 +14,8 @@
 #include <type_traits> // スライス機能のために追加
 #include <fstream>
 #include <cstring>
-using Index = long long;
+
+using Index = int;
 
 // --- スライス用ヘルパー型 ---
